+++ mintapi/api.py
@@ -1,6 +1,7 @@
 import json
 import random
 import time
+import re
 
 try:
     from StringIO import StringIO  # Python 2
@@ -33,12 +34,20 @@ def assert_pd():
             'please pip install pandas'
         )
 
+
 DATE_FIELDS = [
     'addAccountDate',
     'closeDate',
     'fiLastUpdated',
     'lastUpdated',
 ]
+
+MINT_ROOT_URL = 'https://mint.intuit.com'
+MINT_ACCOUNTS_URL = 'https://accounts.intuit.com'
+
+
+class MintException(Exception):
+    pass
 
 
 class MintHTTPSAdapter(HTTPAdapter):
@@ -52,16 +61,17 @@ class Mint(requests.Session):
     request_id = 42  # magic number? random number?
     token = None
 
-    def __init__(self, email=None, password=None):
+    def __init__(self, email=None, password=None, ius_session=None, thx_guid=None):
         requests.Session.__init__(self)
+        self.headers.update({'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9) AppleWebKit/537.71 (KHTML, like Gecko) Version/7.0 Safari/537.71'})
         self.mount('https://', MintHTTPSAdapter())
         if email and password:
-            self.login_and_get_token(email, password)
+            self.login_and_get_token(email, password, ius_session, thx_guid)
 
     @classmethod
-    def create(cls, email, password):  # {{{
+    def create(cls, email, password, ius_session=None, thx_guid=None):  # {{{
         mint = Mint()
-        mint.login_and_get_token(email, password)
+        mint.login_and_get_token(email, password, ius_session, thx_guid)
         return mint
 
     @classmethod
@@ -100,42 +110,97 @@ class Mint(requests.Session):
                                (url, result.status_code))
         if expected_content_type is not None:
             content_type = result.headers.get('content-type', '')
-            if not content_type.startswith(expected_content_type):
+            if not re.match(expected_content_type, content_type):
                 raise RuntimeError(
                     'Error requesting %r, content type %r does not match %r' %
                     (url, content_type, expected_content_type))
         return result
 
-    def login_and_get_token(self, email, password):  # {{{
+    def login_and_get_token(self, email, password, ius_session, thx_guid):  # {{{
         # 0: Check to see if we're already logged in.
         if self.token is not None:
             return
 
         # 1: Login.
-        login_url = 'https://wwws.mint.com/login.event?task=L'
+        login_url = '{}/login.event'.format(MINT_ROOT_URL)
         try:
             self.request_and_check(login_url)
         except RuntimeError:
-            raise Exception('Failed to load Mint login page')
+            raise MintException('Failed to load Mint login page')
 
-        data = {'username': email}
-        response = self.post('https://wwws.mint.com/getUserPod.xevent',
-                             data=data, headers=self.json_headers).text
+        data = {'username': email, 'password': password}
 
-        data = {'username': email, 'password': password, 'task': 'L',
+        # Extract ius_token/thx_guid using browser if not provided manually
+        if not ius_session:
+            session_cookies = self.get_session_cookies(**data)
+        else:
+            session_cookies = {
+                'ius_session': ius_session,
+                'thx_guid': thx_guid
+            }
+        self.cookies.update(session_cookies)
+
+        self.get('https://pf.intuit.com/fp/tags?js=0&org_id=v60nf4oj&session_id=' + self.cookies['ius_session'])
+
+        response = self.post('{}/access_client/sign_in'.format(MINT_ACCOUNTS_URL),
+                             json=data, headers=self.json_headers).text
+
+        json_response = json.loads(response)
+        if json_response.get('action') == 'CHALLENGE':
+            raise MintException('Challenge required, please log in to Mint.com manually and complete the captcha.')
+
+        if json_response.get('responseCode') == 'INVALID_CREDENTIALS':
+            raise MintException('Username/Password is incorrect.  Please verify and try again.')
+
+        data = {'clientType': 'Mint', 'authid': json_response['iamTicket']['userId']}
+        self.post('{}/getUserPod.xevent'.format(MINT_ROOT_URL),
+                  data=data, headers=self.json_headers)
+
+        data = {'task': 'L',
                 'browser': 'firefox', 'browserVersion': '27', 'os': 'linux'}
-        response = self.post('https://wwws.mint.com/loginUserSubmit.xevent',
+        response = self.post('{}/loginUserSubmit.xevent'.format(MINT_ROOT_URL),
                              data=data, headers=self.json_headers).text
 
         if 'token' not in response:
-            raise Exception('Mint.com login failed[1]')
+            raise MintException('Mint.com login failed[1]')
 
         response = json.loads(response)
         if not response['sUser']['token']:
-            raise Exception('Mint.com login failed[2]')
+            raise MintException('Mint.com login failed[2]')
 
         # 2: Grab token.
         self.token = response['sUser']['token']
+
+    def get_session_cookies(self, username, password):
+        try:
+            from selenium import webdriver
+            driver = webdriver.Chrome()
+        except Exception as e:
+            raise RuntimeError("ius_session not specified, and was unable to load "
+                               "the chromedriver selenium plugin. Please ensure "
+                               "that the `selenium` and `chromedriver` packages "
+                               "are installed.\n\nThe original error message was: " +
+                               (e.args[0] if len(e.args) > 0 else 'No error message found.'))
+
+        driver.get("https://www.mint.com")
+        driver.implicitly_wait(20)  # seconds
+        driver.find_element_by_link_text("Log In").click()
+
+        driver.find_element_by_id("ius-userid").send_keys(username)
+        driver.find_element_by_id("ius-password").send_keys(password)
+        driver.find_element_by_id("ius-sign-in-submit-btn").submit()
+
+        # Wait until logged in, just in case we need to deal with MFA.
+        while not driver.current_url.startswith('https://mint.intuit.com/overview.event'):
+            time.sleep(1)
+
+        try:
+            return {
+                'ius_session': driver.get_cookie('ius_session')['value'],
+                'thx_guid': driver.get_cookie('thx_guid')['value']
+            }
+        finally:
+            driver.close()
 
     def get_accounts(self, get_detail=False):  # {{{
         # Issue service request.
@@ -162,13 +227,12 @@ class Mint(requests.Session):
         }
 
         data = {'input': json.dumps([input])}
-        account_data_url = ('https://wwws.mint.com/bundledServiceController.'
-                            'xevent?legacy=false&token=' + self.token)
+        account_data_url = '{}/bundledServiceController.xevent?legacy=false&token={}'.format(MINT_ROOT_URL, self.token)
         response = self.post(account_data_url, data=data,
                              headers=self.json_headers).text
         self.request_id = self.request_id + 1
         if req_id not in response:
-            raise Exception('Could not parse account data: ' + response)
+            raise MintException('Could not parse account data: ' + response)
 
         # Parse the request
         response = json.loads(response)
@@ -191,8 +255,7 @@ class Mint(requests.Session):
         return accounts
 
     def set_user_property(self, name, value):
-        url = ('https://wwws.mint.com/bundledServiceController.xevent?' +
-               'legacy=false&token=' + self.token)
+        url = '{}/bundledServiceController.xevent?legacy=false&token={}'.format(MINT_ROOT_URL, self.token)
         req_id = str(self.request_id)
         self.request_id += 1
         result = self.post(
@@ -204,10 +267,10 @@ class Mint(requests.Session):
                                         'id': req_id}])},
             headers=self.json_headers)
         if result.status_code != 200:
-            raise Exception('Received HTTP error %d' % result.status_code)
+            raise MintException('Received HTTP error %d' % result.status_code)
         response = result.text
         if req_id not in response:
-            raise Exception("Could not parse response to set_user_property")
+            raise MintException("Could not parse response to set_user_property")
 
     def _dateconvert(self, dateraw):
         # Converts dates from json data
@@ -237,15 +300,13 @@ class Mint(requests.Session):
         """
 
         # Warning: This is a global property for the user that we are changing.
-        self.set_user_property('hide_duplicates',
-                               'T' if skip_duplicates else 'F')
+        self.set_user_property('hide_duplicates', 'T' if skip_duplicates else 'F')
 
         # Converts the start date into datetime format - must be mm/dd/yy
         try:
             start_date = datetime.strptime(start_date, '%m/%d/%y')
         except:
             start_date = None
-
         all_txns = []
         offset = 0
         # Mint only returns some of the transactions at once.  To get all of
@@ -255,7 +316,8 @@ class Mint(requests.Session):
             # transactions as well.  Otherwise they are skipped by
             # default.
             url = (
-                'https://wwws.mint.com/getJsonData.xevent?' +
+                MINT_ROOT_URL +
+                '/getJsonData.xevent?' +
                 'queryNew=&offset={offset}&comparableType=8&' +
                 'rnd={rnd}&{query_options}').format(
                     offset=offset,
@@ -265,18 +327,13 @@ class Mint(requests.Session):
                         else 'task=transactions,txnfilters&filterType=cash'))
             result = self.request_and_check(
                 url, headers=self.json_headers,
-                expected_content_type='application/json')
+                expected_content_type='text/json|application/json')
             data = json.loads(result.text)
             txns = data['set'][0].get('data', [])
-            df = pd.DataFrame(txns)
             if start_date:
-                dates = list(df['odate'])
-                last_dt = self._dateconvert(dates[-1])
+                last_dt = self._dateconvert(txns[-1]['odate'])
                 if last_dt < start_date:
-                    keep_txns = []
-                    for item in txns:
-                        if self._dateconvert(item['odate']) >= start_date:
-                            keep_txns.append(item)
+                    keep_txns = [t for t in txns if self._dateconvert(t['odate']) >= start_date]
                     all_txns.extend(keep_txns)
                     break
             if not txns:
@@ -356,8 +413,7 @@ class Mint(requests.Session):
         result = self.request_and_check(
             url,
             headers=self.headers,
-            expected_content_type='text/csv'
-            )
+            expected_content_type='text/csv')
         return result.content
 
     def get_net_worth(self, account_data=None):
@@ -380,12 +436,12 @@ class Mint(requests.Session):
                 net_worth += current_balance
         return net_worth
 
-    def get_transactions(self, start_date=None, end_date=None):
+    def get_transactions(self, start_date=None, end_date=None, include_investment=None):
         """Returns the transaction data as a Pandas DataFrame.
         """
         assert_pd()
         s = StringIO(self.get_transactions_csv(start_date=start_date,
-                                               end_date=end_date))
+                                               end_date=end_date,include_investment=include_investment))
         s.seek(0)
         df = pd.read_csv(s, parse_dates=['Date'])
         df.columns = [c.lower().replace(' ', '_') for c in df.columns]
@@ -399,11 +455,11 @@ class Mint(requests.Session):
         # and parsing the HTML snippet :(
         for account in accounts:
             headers = self.json_headers
-            headers['Referer'] = ('https://wwws.mint.com/transaction.event?'
+            headers['Referer'] = ('{}/transaction.event?'.format(MINT_ROOT_URL) +
                                   'accountId=' + str(account['id']))
 
-            list_txn_url = ('https://wwws.mint.com/listTransaction.xevent?'
-                            'accountId=' + str(account['id']) + '&queryNew=&'
+            list_txn_url = ('{}/listTransaction.xevent?'.format(MINT_ROOT_URL) +
+                            'accountId=' + str(account['id']) + '&queryNew=&' +
                             'offset=0&comparableType=8&acctChanged=T&rnd=' +
                             Mint.get_rnd())
 
@@ -466,22 +522,19 @@ class Mint(requests.Session):
             }])
         }
 
-        cat_url = ('https://wwws.mint.com/bundledServiceController.xevent'
-                   '?legacy=false&token=' + self.token)
+        cat_url = '{}/bundledServiceController.xevent?legacy=false&token={}'.format(MINT_ROOT_URL, self.token)
         response = self.post(cat_url, data=data,
                              headers=self.json_headers).text
         self.request_id = self.request_id + 1
         if req_id not in response:
-            raise Exception('Could not parse category data: "' +
-                            response + '"')
+            raise MintException('Could not parse category data: "' +
+                                response + '"')
         response = json.loads(response)
         response = response['response'][req_id]['response']
 
         # Build category list
         categories = {}
         for category in response['allCategories']:
-            if category['parentId'] == 0:
-                continue
             categories[category['id']] = category
 
         return categories
@@ -500,7 +553,7 @@ class Mint(requests.Session):
         last_year = (str(last_year.month).zfill(2) +
                      '/01/' + str(last_year.year))
         response = json.loads(self.get(
-            'https://wwws.mint.com/getBudget.xevent?startDate=' + last_year +
+            MINT_ROOT_URL + '/getBudget.xevent?startDate=' + last_year +
             '&endDate=' + this_month + '&rnd=' + Mint.get_rnd(),
             headers=self.json_headers
         ).text)
@@ -542,15 +595,12 @@ class Mint(requests.Session):
 
     def initiate_account_refresh(self):
         # Submit refresh request.
-        data = {
-            'token': self.token
-        }
-        self.post('https://wwws.mint.com/refreshFILogins.xevent',
-                  data=data, headers=self.json_headers)
+        data = {'token': self.token}
+        self.post('{}/refreshFILogins.xevent'.format(MINT_ROOT_URL), data=data, headers=self.json_headers)
 
 
-def get_accounts(email, password, get_detail=False):
-    mint = Mint.create(email, password)
+def get_accounts(email, password, get_detail=False, ius_session=None):
+    mint = Mint.create(email, password, ius_session=ius_session)
     return mint.get_accounts(get_detail=get_detail)
 
 
@@ -560,11 +610,18 @@ def get_net_worth(email, password):
     return mint.get_net_worth(account_data)
 
 
-def make_accounts_presentable(accounts):
+
+def make_accounts_presentable(accounts, presentable_format='EXCEL'):
+    formatter = {
+        'DATE': '%Y-%m-%d',
+        'ISO8601': '%Y-%m-%dT%H:%M:%SZ',
+        'EXCEL': '%Y-%m-%d %H:%M:%S',
+    }[presentable_format]
+
     for account in accounts:
         for k, v in account.items():
             if isinstance(v, datetime):
-                account[k] = repr(v)
+                account[k] = v.strftime(formatter)
     return accounts
 
 
@@ -609,6 +666,17 @@ def main():
                          'implies --accounts)')
     cmdline.add_argument('--transactions', '-t', action='store_true',
                          default=False, help='Retrieve transactions')
+    cmdline.add_argument('--extended-transactions', action='store_true', default=False,
+                         help='Retrieve transactions with extra information and arguments')
+    cmdline.add_argument('--start-date', nargs='?', default=None,
+                         help='Earliest date for transactions to be retrieved from. Used with --extended-transactions. Format: mm/dd/yy')
+    cmdline.add_argument('--include-investment', action='store_true', default=False,
+                         help='Used with --extended-transactions')
+    cmdline.add_argument('--skip-duplicates', action='store_true', default=False,
+                         help='Used with --extended-transactions')
+# Displayed to the user as a postive switch, but processed back here as a negative
+    cmdline.add_argument('--show-pending', action='store_false', default=True,
+                         help='Exclude pending transactions from being retrieved. Used with --extended-transactions')
     cmdline.add_argument('--filename', '-f', help='write results to file. can '
                          'be {csv,json} format. default is to write to '
                          'stdout.')
@@ -619,6 +687,8 @@ def main():
                          help='Start date for transactions')
     cmdline.add_argument('--end', '-ed', nargs='?', dest='end_date',
                          help='End date for transactions')
+    cmdline.add_argument('--session', help='ius_session cookie')
+    cmdline.add_argument('--thx_guid', help='thx_guid cookie')
 
     options = cmdline.parse_args()
 
@@ -626,8 +696,10 @@ def main():
         cmdline.error('--keyring can only be used if the `keyring` '
                       'library is installed.')
 
-    try:
+    try:  # python 2.x
         from __builtin__ import raw_input as input
+    except ImportError:  # python 3
+        from builtins import input
     except NameError:
         pass
 
@@ -655,11 +727,11 @@ def main():
     if options.accounts_ext:
         options.accounts = True
 
-    if not any([options.accounts, options.budgets, options.transactions,
+    if not any([options.accounts, options.budgets, options.transactions, options.extended_transactions,
                 options.net_worth]):
         options.accounts = True
 
-    mint = Mint.create(email, password)
+    mint = Mint.create(email, password, ius_session=options.session, thx_guid=options.thx_guid)
 
     data = None
     if options.accounts and options.budgets:
@@ -690,12 +762,18 @@ def main():
             data = None
     elif options.transactions:
         data = mint.get_transactions(start_date=options.start_date,
-                                     end_date=options.end_date)
+                                     end_date=options.end_date,include_investment=options.include_investment)
+    elif options.extended_transactions:
+        data = mint.get_detailed_transactions(
+            start_date=options.start_date,
+            include_investment=options.include_investment,
+            remove_pending=options.show_pending,
+            skip_duplicates=options.skip_duplicates)
     elif options.net_worth:
         data = mint.get_net_worth()
 
     # output the data
-    if options.transactions:
+    if options.transactions or options.extended_transactions:
         if options.filename is None:
             print(data.to_json(orient='records'))
         elif options.filename.endswith('.csv'):
@@ -712,6 +790,7 @@ def main():
                 json.dump(data, f, indent=2)
         else:
             raise ValueError('file type must be json for non-transaction data')
+
 
 if __name__ == '__main__':
     main()